permission ui complete

+++ huaimi-database/src/main/java/com/ieng/huaimi/database/service/impl/PermissionServiceImpl.java
@@ -1,5 +1,7 @@
 package com.ieng.huaimi.database.service.impl;
 
+import com.ieng.huaimi.common.exception.ServiceCode;
+import com.ieng.huaimi.common.exception.ServiceException;
 import com.ieng.huaimi.common.utils.string.StringUtil;
 import com.ieng.huaimi.database.domain.Permission;
 import com.ieng.huaimi.database.mapper.PermissionDao;
@@ -7,7 +9,9 @@ import com.ieng.huaimi.database.service.PermissionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 import tk.mybatis.mapper.entity.Example;
+import tk.mybatis.mapper.weekend.WeekendSqls;
 
 import java.util.*;
 
@@ -37,8 +41,14 @@ public class PermissionServiceImpl implements PermissionService {
     }
 
     @Override
-    public List<Permission> findPermissionTree() {
-        return permissionDao.selectAll();
+    public List<Permission> findPermissionCondition(Permission permission) {
+        Example example = new Example(Permission.class);
+        example.and().andEqualTo("type", permission.getType());
+        example.and().andEqualTo("status", permission.getStatus());
+        if(!StringUtils.isEmpty(permission.getPerms())){
+            example.and().andLike("perms", String.format("%s%s%s", "%", permission.getPerms(), "%"));
+        }
+        return permissionDao.selectByExample(example);
     }
 
     @Override
@@ -67,8 +77,10 @@ public class PermissionServiceImpl implements PermissionService {
     @Transactional
     @Override
     public int delPermission(Long id) {
+        if(id <= 16){
+            throw new ServiceException(ServiceCode.ADMINISTRATOR_FORBID);
+        }
         return permissionDao.deleteByPrimaryKey(id);
     }
-
 
 }
